seperating js, adding animation, fixing style
